+++ core/include/cubos/core/ecs/reflection.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cubos/core/reflection/traits/categorizable.hpp>
 #include <cubos/core/reflection/traits/constructible.hpp>
 #include <cubos/core/reflection/traits/fields.hpp>
 #include <cubos/core/reflection/traits/wrapper.hpp>
@@ -92,6 +93,21 @@ namespace cubos::core::ecs
         TypeBuilder&& withField(std::string name, F T::* pointer) &&
         {
             mFields.addField(std::move(name), pointer);
+            return std::move(*this);
+        }
+
+        /// @brief Adds a category to a type.
+        ///
+        /// Useful for grouping related components close to each other, for example,
+        /// in the entity inspector.
+        ///
+        /// @param category Category name.
+        /// @param priority Priority of a type within its category.
+        /// @return Builder.
+        TypeBuilder&& categorize(std::string category, size_t priority) &&
+        {
+            CUBOS_ASSERT(!mType.has<reflection::CategorizableTrait>());
+            mType.with(reflection::CategorizableTrait(std::move(category), priority));
             return std::move(*this);
         }